Final Organization
- Since renaming NavMeshes seem to break them I'll keep them as they are. But Considering they use the name of the object they belong I'll name all the "Grounds" differently.
- All NavMeshes are in folders for scenes they belong on a NavMesh folder within _scenes.
- Also got rid of all unnecessary files and placed every other temporary and non-temporary files on the correct folders.
- Changed build settings in order to use the revamped main menu UI. Kept old scene under temporary files as a BackUp!
