+++ Assets/Materials/ClickEffectMat.mat.meta
