+++ Assets/Materials/PlaceholderMaterials/DevWaypoint.mat
