+++ Assets/Materials/PlaceholderMaterials/DevWaypoint.mat.meta
