+++ Assets/Miscelaneous.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: fc5b9a30bcaee96418564be4df52616f
-NativeFormatImporter:
+guid: 7f25601432e413d4c8bf6e188f4c928d
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
