--- Assets/_Scenes/InGame_A1/NavMesh-Ground.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cea6063035c597d46b7f74b1dd75455c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
