+++ Assets/_Scenes/NavMeshes/Level_1.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 7f6857f8c3280da9b8a48792bd060d56
-NativeFormatImporter:
+guid: 668003bec5a398d41840777c06da4a56
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
