+++ Assets/_Scenes/NavMeshes/Level_1/NavMesh-GroundLevel1.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: ae003200a98e18f499356d580efea0d3
-folderAsset: yes
-DefaultImporter:
+guid: 1a47e785ca8b47d4da39f24c235de3c0
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
