--- Assets/_Scenes/RedesignStartMenu.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7eaf4a47c55bbdf4cba4c2104076138a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
