+++ Assets/_Scenes/Temporary/StartMenuOld.unity.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: f7bcd9b249027224d8299feac7c70e84
-NativeFormatImporter:
+guid: 681a31293ec0499b992905c45a60356d
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
