+++ ProjectSettings/EditorBuildSettings.asset
@@ -7,7 +7,7 @@ EditorBuildSettings:
   m_Scenes:
   - enabled: 1
     path: Assets/_Scenes/_StartMenu.unity
-    guid: 681a31293ec0499b992905c45a60356d
+    guid: 7eaf4a47c55bbdf4cba4c2104076138a
   - enabled: 1
     path: Assets/_Scenes/Level_1.unity
     guid: 995afc9ae90f0d742b99714f09627faf
